services/rag_loading_service/requirements.txt
kafka-python>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
langchain>=0.0.300
chroma-core>=0.4.0
openai>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0